LPS-93431 Move the upgrade process to the proper place so that we can backport it to 7.1
