+++ modules/apps/dynamic-data-lists/dynamic-data-lists-service/bnd.bnd
@@ -6,7 +6,7 @@ Import-Package:\
 	!org.apache.poi.ss.usermodel.*,\
 	\
 	*
-Liferay-Require-SchemaVersion: 2.1.1
+Liferay-Require-SchemaVersion: 2.1.0
 Liferay-Service: true
 -dsannotations-options: inherit
 -includeresource: @poi-[0-9]*.jar
