+++ modules/apps/dynamic-data-lists/dynamic-data-lists-service/src/main/java/com/liferay/dynamic/data/lists/internal/upgrade/DDLServiceUpgrade.java
@@ -23,7 +23,7 @@ import com.liferay.dynamic.data.lists.internal.upgrade.v2_0_0.util.DDLRecordSetT
 import com.liferay.dynamic.data.lists.internal.upgrade.v2_0_0.util.DDLRecordSetVersionTable;
 import com.liferay.dynamic.data.lists.internal.upgrade.v2_0_0.util.DDLRecordTable;
 import com.liferay.dynamic.data.lists.internal.upgrade.v2_0_0.util.DDLRecordVersionTable;
-import com.liferay.dynamic.data.lists.internal.upgrade.v2_1_1.UpgradeVersionUserId;
+import com.liferay.dynamic.data.lists.internal.upgrade.v1_1_1.UpgradeVersionUserId;
 import com.liferay.dynamic.data.mapping.service.DDMStructureLocalService;
 import com.liferay.portal.kernel.upgrade.BaseUpgradeSQLServerDatetime;
 import com.liferay.portal.upgrade.registry.UpgradeStepRegistrator;
@@ -64,6 +64,8 @@ public class DDLServiceUpgrade implements UpgradeStepRegistrator {
 			new com.liferay.dynamic.data.lists.internal.upgrade.v1_1_0.
 				UpgradeDDLRecordSetVersion(_counterLocalService));
 
+		registry.register("1.1.0", "1.1.1", new UpgradeVersionUserId());
+
 		registry.register(
 			"1.1.0", "2.0.0",
 			new BaseUpgradeSQLServerDatetime(
@@ -76,8 +78,6 @@ public class DDLServiceUpgrade implements UpgradeStepRegistrator {
 			"2.0.0", "2.1.0",
 			new com.liferay.dynamic.data.lists.internal.upgrade.v2_1_0.
 				UpgradeSchema());
-
-		registry.register("2.1.0", "2.1.1", new UpgradeVersionUserId());
 	}
 
 	@Reference
